Renaming 2 files so they start with a capital letter
