+++ frontend/app/components/ProductionGridViewItem.vue
