+++ frontend/app/components/ProductionListViewItem.vue
